backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv
duckdb
polars
requests_cache
pytest
requests_mock
httpx 